feat: added rsc support changes next
