+++ packages/react-native-css-interop/src/runtime/api.ts
@@ -1,1 +1,2 @@
 export * from "./web/api";
+export * from "./web/interopComponentsMap";
