+++ packages/react-native-css-interop/src/runtime/components.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import {
   ActivityIndicator,
   FlatList,
