+++ packages/react-native-css-interop/src/runtime/web/interopComponentsMap.ts
@@ -1,0 +1,6 @@
+import { JSXFunction } from "../../types";
+
+export const interopComponents = new Map<
+  object | string,
+  Parameters<JSXFunction>[0]
+>();
